refactor(deploy-task): renamed numeric representation
